backend/requirements.txt
fastapi[standard]==0.118.2
uvicorn[standard]==0.30.6
websockets==14.1
pydantic>=2.7.0,<3.0.0
pydantic-settings==2.2.1
pydantic[email]
pandas==2.2.2
numpy==2.2.0
python-dotenv==1.0.1
httpx==0.28.1
aiohttp==3.10.11
requests==2.32.3
slowapi==0.1.9
pytest==8.3.3
pytest-asyncio==0.24.0
ruff==0.8.4
mypy==1.13.0
